reactRouterEstablishmentForHeadNav
Jessrey added react router functionality for easy traversing of the content that is to be added. Please note that there are comments that tell you where to add your component to the code.

please add components to the component folder and the main pages at the src file directly. OR YOU CAN CREATE A PAGE FOLDER IF YOU WISH JUST PLEASE INFORM THE GROUP.